backend/app/models/department.py
from typing import List, TYPE_CHECKING

from sqlalchemy import String, Text
from sqlalchemy.orm import Mapped, mapped_column, relationship
from sqlalchemy.dialects.postgresql import UUID

from .base import Base

if TYPE_CHECKING:
    from .user import User


class Department(Base):
    __tablename__ = "departments"

    id: Mapped[UUID] = mapped_column(UUID(as_uuid=True), primary_key=True)
    name: Mapped[str] = mapped_column(String(128))
    abbreviation: Mapped[str] = mapped_column(String(8))
    description: Mapped[str] = mapped_column(Text())

    users: Mapped[List["User"]] = relationship()
